requirements.txt
# Python dependencies
torch>=2.0.0
torch-geometric>=2.3.0
torch-scatter>=2.1.0
torch-sparse>=0.6.0
torch-cluster>=1.6.0

# Data processing
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
umap-learn>=0.5.0
plotly>=5.0.0

# Deep learning utilities
tqdm>=4.62.0
tensorboard>=2.8.0
wandb>=0.12.0

# Configuration
pyyaml>=5.4.0
omegaconf>=2.1.0

# Notebook support
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=7.6.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0

# Utilities
h5py>=3.1.0
networkx>=2.6.0
python-louvain>=0.16
